fix: keep node-only logger out of browser bundle
`client.ts` is bundled into the browser and into the client's service
worker, and reached `logger` through `./notification` →
`UserNotifications`. `logger` reads `process.env.LOG_LEVEL` at module
scope, which is a ReferenceError where there is no `process` — service
worker registration then fails outright and the app never leaves
"Installing application...".

Move the dispatch helper, the only part that logs, into
`notification/dispatch.ts`, and take `TriggerEvent` from the leaf module
so the browser entry stops pulling the index in.
