+++ packages/commons/src/client.ts
@@ -23,7 +23,7 @@ export * from './url'
 export * from './events/serializers/user/serializer'
 export * from './events/serializers/date/serializer'
 export * from './events/mocks.test.utils'
-export { TriggerEvent } from './notification'
+export { TriggerEvent } from './notification/UserNotifications'
 export * from './application-config'
 export * from './icons'
 export { findPendingCorrectionAction } from './events/state/flags'
